feat: useTimer hook 분리 (#34)
